+++ tests/Executor/DeferredFieldsTest.php
@@ -592,10 +592,10 @@ class DeferredFieldsTest extends TestCase
             }
         ');
 
-        $author1 = ['name' => 'John'/*, 'bestFriend' => ['name' => 'Dirk']*/];
-        $author2 = ['name' => 'Jane'/*, 'bestFriend' => ['name' => 'Joe']*/];
-        $author3 = ['name' => 'Joe'/*, 'bestFriend' => ['name' => 'Jane']*/];
-        $author4 = ['name' => 'Dirk'/*, 'bestFriend' => ['name' => 'John']*/];
+        $author1 = ['name' => 'John'/* , 'bestFriend' => ['name' => 'Dirk'] */];
+        $author2 = ['name' => 'Jane'/* , 'bestFriend' => ['name' => 'Joe'] */];
+        $author3 = ['name' => 'Joe'/* , 'bestFriend' => ['name' => 'Jane'] */];
+        $author4 = ['name' => 'Dirk'/* , 'bestFriend' => ['name' => 'John'] */];
 
         $story1 = ['title' => 'Story #8', 'author' => $author1];
         $story2 = ['title' => 'Story #3', 'author' => $author3];